DescaleTarget: Add shifting, cross-conversion support (#7)
* DescaleTarget: Add shifting, cross-conversion support

* DescaleTarget: Set `line_mask` to False by default if cross-conversion

* DescaleTarget: Move crossconv shift logic to own method

Deduplication, also allows us to call upon this later in case we need to check it.

* DescaleTarget: New staticmethods for crossconv shifts

Not entirely sure what to do with this. For now, they're staticmethods so the user can call them themselves. According to Zewia, these are the only shifts he's seen out in the wild.

* DescaleTarget: Update shift handling

[This commit](https://github.com/Irrational-Encoding-Wizardry/vs-kernels/commit/8705da5dbcb734694c909d98b24b798eac4f8f09) adds proper field-based descaling support, so some code here is no longer necessary.

* DescaleTarget: Fix `fields=None`

oops

* DescaleTarget: Re-add line_mask with cross-conversion

Was initially removed because it's not safe with cross-conversion, as it tended to miss large parts of the lineart. This has been re-added now, and the edgemask simply gets inflated a ton.  Users who don't want it to mask can simply pass `line_mask=False`.

* DescaleTarget: Update docstring on `fields`

* DescaleTarget: `shift_*` => `shift`

This was different before to allow per-field shifting, but `vskernels` handles this automatically now.

* DescaleTarget: `fields` => `field_based`

* DescaleTarget: Add missing class name to ValueErrors