DescaleTarget: Add borders to edgemask if border_handling, remove vsmask
This was added after the regular linemask processing so the user can still pass their own linemask without having to worry about the borders being gone.

2 pixels are chosen because those are *typically* the most I see for dirty edges. The user can simply pass their own mask if they really need to increase that.

This commit also deprecates the old vsmask package. vsmasktools is in the requirements and vsmask has not been updated in a very long time, so there's little reason to still keep it around.