DescaleTarget: Add dynamic border radius for border_handling
Uses the newly added kernel_radius from vskernels if available. If not implemented (weird kernel usage) or can't be found (older version), it falls back to a radius of 2.

Furthermore, the radius gets expanded depending on the scale ratio. As discussed with @arch1t3cht, high scale ratios (such as 540 -> 1080) can make the border appear larger (as it's the source image that matters for the kernel radius). We try to compensate for that here, with a bit of headroom for fractional resolutions. Right now, an extra pixel gets added if the target size is ~708 or lower.