DescaleTarget: better bord mask handling, other changes
I moved the border mask creation to another method, as it must be applied to both the credit mask and the line mask. It's unfortunate, because this means it will harm scrolling credits a little bit, but usually those show up in scenes that shouldn't be rescaled anyway (think black screen with scrolling credits).

I also added additional scaling ratio logic. The kernel size only matters in relation to the original image, so high scale ratios (like 2, i.e. 540 -> 1080) would have thicker borders that requires a bigger mask.

Finally, I moved the Limiter calls for the credit/line mask to the moment they get dithered. It seems to just play nicer with the border mask, and is probably safer no matter what.