DescaleTarget: Add border_handling (#10)
* DescaleTarget: Add border_handling params

* DescaleTarget: Add borders to edgemask if border_handling, remove vsmask

This was added after the regular linemask processing so the user can still pass their own linemask without having to worry about the borders being gone.

2 pixels are chosen because those are *typically* the most I see for dirty edges. The user can simply pass their own mask if they really need to increase that.

This commit also deprecates the old vsmask package. vsmasktools is in the requirements and vsmask has not been updated in a very long time, so there's little reason to still keep it around.

* Fix Formatting (again)

* descale: Fix imports

* DescaleTarget: Add dynamic border radius for border_handling

Uses the newly added kernel_radius from vskernels if available. If not implemented (weird kernel usage) or can't be found (older version), it falls back to a radius of 2.

Furthermore, the radius gets expanded depending on the scale ratio. As discussed with @arch1t3cht, high scale ratios (such as 540 -> 1080) can make the border appear larger (as it's the source image that matters for the kernel radius). We try to compensate for that here, with a bit of headroom for fractional resolutions. Right now, an extra pixel gets added if the target size is ~708 or lower.

* DescaleTarget: better bord mask handling, other changes

I moved the border mask creation to another method, as it must be applied to both the credit mask and the line mask. It's unfortunate, because this means it will harm scrolling credits a little bit, but usually those show up in scenes that shouldn't be rescaled anyway (think black screen with scrolling credits).

I also added additional scaling ratio logic. The kernel size only matters in relation to the original image, so high scale ratios (like 2, i.e. 540 -> 1080) would have thicker borders that requires a bigger mask.

Finally, I moved the Limiter calls for the credit/line mask to the moment they get dithered. It seems to just play nicer with the border mask, and is probably safer no matter what.

* DescaleTarget: Move border_handling masking to upscale method

* DescaleTarget: new param: border_radius

* DescaleTarget: Reintroduce dirty borders

* DescaleTarget: Fix imports from latest commit

* DescaleTarget: Improve border masking

* DescaleTarget: Add repeat border handling

* DescaleTarget: Fix AttributeError with frac

* DescaleTarget: Fix accounting for frac descales

* DescaleTarget: Improve border equation

* DescaleTarget: Remove bordermask getter

* DescaleTarget: Fix border behavior and optimize merging

* DescaleTarget: remove useless param

---------

Co-authored-by: Vodes <48185123+Vodes@users.noreply.github.com>